third_party/blink/web_tests/fast/dom/shadow/tooltips-in-shadow-expected.txt

Test for Case1: Non-Nested
PASS harness.expected is testRunner.tooltipText
Test for Case 2: Nested
PASS harness.expected is testRunner.tooltipText
PASS successfullyParsed is true

TEST COMPLETE
